requirements.txt
### `requirements.txt`

plaintext
fastapi
python-dotenv
openai
uvicorn